test/requirements.txt
pytest
redis
kafka-python
pylightxl
kubernetes
pymysql